Image Bundling Part 1
The last major feature is a way to identify bundles of images that are used for materials, and representing them as a single entry in the interface.  The backend work has been mostly completed, now I need to add the string matching and some additional interface tweaks to support it.