Updated File Data Creation UI
The UI was updated to provide specifics on what exactly is added to the file when the button is pressed and to remind the user not to delete these things when they are added.

Info was also added about Compositor Output nodes even though the addon doesn't add them yet - this will happen once Image Format Settings are added fully.